+++ src/Defaults/IceHawkConfig.php
@@ -17,8 +17,8 @@ use IceHawk\IceHawk\Defaults\Traits\DefaultEventSubscribing;
 use IceHawk\IceHawk\Defaults\Traits\DefaultFinalReadResponding;
 use IceHawk\IceHawk\Defaults\Traits\DefaultFinalWriteResponding;
 use IceHawk\IceHawk\Defaults\Traits\DefaultReadRouting;
+use IceHawk\IceHawk\Defaults\Traits\DefaultRequestBypassing;
 use IceHawk\IceHawk\Defaults\Traits\DefaultRequestInfoProviding;
-use IceHawk\IceHawk\Defaults\Traits\DefaultRedirectRouting;
 use IceHawk\IceHawk\Defaults\Traits\DefaultWriteRouting;
 use IceHawk\IceHawk\Interfaces\ConfiguresIceHawk;
 
@@ -30,7 +30,7 @@ class IceHawkConfig implements ConfiguresIceHawk
 {
 	use DefaultReadRouting;
 	use DefaultWriteRouting;
-	use DefaultRedirectRouting;
+	use DefaultRequestBypassing;
 	use DefaultEventSubscribing;
 	use DefaultRequestInfoProviding;
 	use DefaultFinalReadResponding;